Moved Clients into subdirectory to clear up file structure
This commit moves all of the clients and the client mocks into a
subdirectory to allow for a cleaner file structure for the project.
The alias of them within the tests is for the Clients directory
themselves to not cause confusion between what we are mocking with in
the Services tests. Having Clients.HTTPMock or Clients.LambdaMock
aliased made it appear is if we were mocking out the HTTP Services
themselves and could lead to confusion later down the line.