Add runtime config mechanism
This runtime config isn't used when running `mix run` but is used
as part of `mix release`. This means that development config is
separate to Cosmos config which is unfortunate but probably the least
bad outcome.

I've introduced a safety check so that whenever a Cosmos config
is not set in the environment, the application will fail to start.
This prevents the case where a config value is mysteriously empty and
fails more obviously.