Use full atom for gen_server metric name
Previously this was taking just the last element of the atom if it
contained any dots (.), for example "Belfrage.Worker" would become
"Worker". This starts to become a problem when things such as
"ExMetrics.Statsd.Worker" get involved, so now it uses the full name,
drops any leading "Elixir" prefix, and converts dots to underscores.