Retrieve pool status once when reporting metrics
Before this change the process that reports metrics for HTTP pools retrieved
status of each pool twice: to get the number of workers and the number of
overflow workers separately, which is unnecessary.

This commit also renames the process to PoolObserver and contains some
refactoring which hopefully makes it more explicit, clearer and more readable.