Add .tool-versions with Erlang/Elixir versions to be used
Having `.tool-versions` should simplify installation and upgrading. README
already recommends using asdf, so we might as well specify the versions that
should be used in the format that asdf understands. It also allows not to have
these versions installed globally. Installation instructions in README were
updated accordingly.

We've recently updated Erlang/Elixir versions in #859, so using the versions we
upgraded to.