Silence log output when running tests
Currently when tests run, log output is displayed which is quite noisy. Passing
`capture: true` to `ExUnit.start` makes `ExUnit` capture log output and display
it on test failures only, which is what we want: we only want to see if it
something goes wrong so that we could debug test failures.