Disable 'waiting' queue in poolboy HTTP pools
We use MachineGun to make HTTP requests to origins and it maintains pools of
workers that actually make those requests using poolboy. When poolboy is asked
for a worker and there are none available, by default, poolboy adds the request
into a queue and attempts to process it later when a worker becomes available.
When new requests come in faster than poolboy can process them, the queue grows
and can eventually become too expensive to manage, which is why we'd like to
disable this behaviour and instead make poolboy return an error if there are no
workers available in the pool. Poolboy allows to do that, but it requires a
change to the way MachineGun uses poolboy, which is implemented in the fork.