Set pool_timeout to a high value for HTTP pools
MachineGun uses poolboy to maintain pools of workers that make HTTP requests.
pool_timeout is the amount of time that poolboy waits when attempting to check
out a worker from a pool.

In theory it's useful to have this timeout set so that we don't wait
indefinitely. However, in practice, to handle this timeout poolboy sends itself
an extra `cancel_waiting` message which seems to be expensive to process and
causes poolboy's mailbox to grow under high load. When the load subsides, it
takes poolboy a long time to process the backlog of messages and so it
effectively remains unusable for new requests until that is finished.

We're setting this timeout to a high enough value to effectively disable
it and address the issue. We've done load testing to confirm that it
works. We'll now rely on cowboy's timeout to ensure that a client's
request handles is not blocked indefinitely at the point of attempting
to checkout out a worker from a poolboy pool.

We're setting it to 5 minutes which is the default `inactivity_timeout`
in cowboy. This should never be reached in reality. This seems better
than setting the timeout to `:infinity` as it provides a guarantee that
we'll never keep waiting for poolboy forever if something goes really
wrong.