Check presence of catch-all route before defining it in `no_match`
A catch-all GET route is defnied by `handle_proxy_pass "/*any"` for test
environment and this causes a compilation warning when another one is
defined by `no_match` (routes are effectively Elixir functions and so
the compiler complains about two identical function clauses). This
change will ensure that `no_match` defines it only on live: we'll check
that the route hasn't been defined before defining it in `no_match`.