Remove redundant private function in Metrics.LatencyMonitor
Having `append_time` as a private function doesn't have much benefits. It's
only used by one other function which already checks that the name of the
checkpoint is in `@valid_checkpoints`, so the logic in `append_time` could as
well be inline in the caller function. Having that logic inline takes less
space and you can see it all at once without having to jump from line to line
following the call stack.