Simplify cleanup in Metrics.LatencyMonitor
Scheduling the next cleanup and doing the actual filtering of requests inline
is more concise and makes all the logic visible at once. It's not complex
enough to justify so many private functions.

Having multiple function heads is helpful when the function is actually called
with different parameters, but using it for breaking down logic into steps
makes it harder to follow the logic.