Handle previously cleaned-up requests in LatencyMonitor
When `LatencyMonitor` processes a `:response_end` checkpoint for a request, the
request may have already been cleaned up. This can happen under high load when
`LatencyMonitor` can't keep up with the number of requests it's asked to handle
and its process mailbox starts to grow. In such scenario processing of a
`:response_end` checkpoint can happen with a significant delay and so by the
time it's processed the request may already have been cleaned-up.

Before this change an exception would be raised because we'd attempt to add the
passed time to a request that's no longer in the state. This commit makes
`LatencyMonitor` ignore such requests.

I considered increasing `@cleanup_ttl` to make sure request timings are kept
for longer and therefore we don't loose metrics in such scenario, but we've
seen during load testing that `LatencyMonitor`'s mailbox size can grow to
hindrends of thousands of messages and so this would still be happening even if
I did increase it.