Track process crashes in production
This commit adds tracking for process crashes in production. It's
implemented as a `Logger` backend which seems to be the established way
of being notified about crashes in a generic way without having to add
instrumentation to individual processes: e.g. this is what tools like
AppSignal do.

The backend is only enabled in production. It intercepts errors with a
`:crash_reason` in metadata which is set by `Logger` when e.g. a
`GenServer` crashes:
https://hexdocs.pm/logger/Logger.html#module-metadata and increments a
metric.

The test attempts to verify that it works as a whole by causing crashes
of a test process and ensuring that an appropriate metric is recorded.

The test uses `Mox` to set an expectation on a mock (`EventMock`) that's
called from a different process (`CrashTracker` logger backend), so it
has use `Mox` in global mode and can't be async.

The test needs to catch any errors that happen in `CrashTracker` and
because that process is started indirectly via `Logger`, the test has to
monitor the tracker, check its termination status and fail if it
terminated abnormally (which would happen if e.g. `Mox` expectation
doesn't match).

`Logger` docs mention three types of things that can cause a crash:
exception, exit and uncaught throw, so the test verifies them all, even
though I'm not sure if all can actually happen in our case.