Don't automatically compile routefile mocks
We're curently re-compiling these routefile mocks in `test_helper.exs`
while setting a `production_environment` config value to specific values
to ensure that the routes in those mocks that are only supposed to be
defined in those environments are indeed defined. This causes a warning
because those modules are already compiled (they are compiled by default
with all other modules in the app).

Changing the extension of the files with those modules to `.exs` ensures
that they are not compiled by default and so we end-up compiling them
once in `test_helper.exs`.