Don't recompile routefile when running tests
It's necessary to recompile the routefile when starting test instances
of belfrage because the code for those instances is compiled with
`production_environment` set to "live", however it's not necessary when
running tests: the code is compiled with `production_environment` set to
"test" when tests are run.

This fixes the warning that is displayed by the compiler when the
routefile is recompiled when running tests.