Isolate and simplify the test for RouteMaster.no_match
When testing RouteMaster we're interested in testing the routes that it
defines with its macros. It's easier to do that with an isolated
routefile, so we can define one for a test and not rely on a giant
shared routefile.

Also removed redundant setup (e.g. we don't always need BBC headers).
We're testing RouteMaster here, not the entire app, so we should only
setup the bare minimum to make it clear what RouteMaster depends on.