Check env when searching for existing catch-all GET route
When we define a proxy-pass route we check the production environment
and don't define it if the current environment doesn't match the one
passed in `only_on` argument. However, we add it to `@routes` in any
case.

We rely on `@routes` in `no_match`: we don't define a catch-all GET
route if there's one already on `@routes` to avoid having duplicate
route matchers. The previous behaviour meant that `no_match` never
defined the catch-all GET route because there'd always be the one
defined by `handle_proxy_pass`. To fix this we also need to check the
environment of the existing route: if it doesn't match the current env,
it means that the route hasn't actually been defined yet and so we need
to define it in `no_match`.