Call latency checkpoints in Services.HTTP instead of Clients.HTTP
Doing this in `Services.HTTP` makes more sense because `Clients.HTTP` is
used for other things as well as making HTTP requests to origins (e.g.
fetching a cached page from S3).

I assume that it was done in `Clients.HTTP` to make it as close to
calling MachineGun as possible to increase accuracy. I don't think that
doing this in `Services.HTTP` make it significantly less accurate: the
only thing that `Clients.HTTP.execute` does is adapting a
`MachineGun.Response` struct to `Clients.HTTP.Response` struct, which
should be a very quick operation. Besides, if we were serious about
accuracy we'd need to implement a way to track how long a request that
`:gun` performs under the hood takes without all the extra stuff that
`MachineGun` does like pooling.

This also fixes the issue with overwritting latency checkpoints when
fetching a fallback from the distributed cache (which uses
`Clients.HTTP`). It means that fetching a fallback will now be part of
internal response latency, which I think makes sense: it's an integral
part of Belfrage and if it's inefficient or becomes slow under high load
we would like to see that (and not attribute it to "external latency").