Consider fetching early response from cache as internal latency
When Belfrage is able to fetch a fresh response from local cache, it
doesn't send a request to the origin at all. Arguably, fetching from
local cache is an intergral part of Belfrage and should therefore be
considered "internal latency" too: if it's inefficient or becomes slow
under load, we'd like to be able to see that in our metrics.

This commit removes latency tracking from `Cache.Local` and that stops
tracking fetching from cache as "external latency". However because we
don't send a request to an origin, we need some other way to record
where processing the request stops and processing the response begins.
This commit adds a new checkpoint for early responses that is recorded
in `Belfrage` after an early response is fetched from cache and response
processing starts. This seems to be appropriate as that's when response
processing begins in Belfrage. If we wanted to have accurate metrics for
e.g. just fetching from the local cache, we could add that separately
later.

Incidentally, this also fixes the issue with overwriting latency
checkpoints when we attempt to fetch a fallback from the local cache.

I had to make `BelfrageTest` synchronous because it starts
`LatencyMonitor` process which another process can attempt to start as
well. A proper fix would provavly be to always start `LatencyMonitor` in
test env instead, but that needs to be looked into separately.