Refactor LatencyMonitorTest
Updated the test to use LatencyMonitor's API to record and retrieve
checkpoints rather than testing it's GenServer callbacks, which doesn't
guarantee that they are actually called.

The way recording the metrics is tested should probably be improved:
starting a UDP server and verifying the metrics in StatsD format is
probably too low-level. I'm leaving this for now though.