Fix flaky MailboxMonitor tests
The tests that verified recording of a loop's mailbox size were flaky
before this change. They'd send a message to the loop and then
run MailboxMonitor assuming that the message sent to the loop would
still be in the loop's mailbox. There was nothing preventing the loop
from processing the message before MailboxMonitor would run though,
which is why those tests would fail sometimes.

I fixed it by suspending the loop process before sending it messages,
which ensures that it doesn't attempt to process them, and so guarantees
that the mailbox size is what we want it to be.

Besides, I made these changes:

* Tests now define its own loop instead of relying on an existing one.
This avoids side effects on other tests that e.g. don't expect the
shared loop to be running.
* Used `start_supervised!` to start the loop process instead of using
`Loops.Supervisor`. This ensures that the loop process is terminated
when the test finishes (before the loop would continue to run).
* Stopped relying on calling `MailboxMonitor.handle_info` to trigger the
metric: it's MailboxMonitor's internal implementation and the fact that
the function works doesn't guarantee that it's actually ever called.
It's better to treat `MailboxMonitor` as a black box if possible. The
tests now start it and wait for the initial `:refresh` message to be
processed.

The test can no longer be async because we're setting Mox expectations
on `Event` that is called from `MailboxMonitor` process, i.e. a
different process to the one the tests run in.