Update remaining MailboxMonitor tests
Updated the tests that verify recording mailbox size for processes to be
similar to the tests that verify that for loops:

* Tests now define its own GenServer inline instead of relying on one in
test/support/mocks. I think this is better because it avoids coupling
between tests (that could re-use the same GenServer) and makes it easier
to see what the GenServer under test does because it's defined right
there.
* Stopped relying on calling `MailboxMonitor.handle_info` to trigger the
metric: it's MailboxMonitor's internal implementation and the fact that
the function works doesn't guarantee that it's actually ever called.
It's better to treat `MailboxMonitor` as a black box if possible. The
tests now start it and wait for the initial `:refresh` message to be
processed.