Start dials in test env
This change is prompted by me trying to fix a flaky test.

I've noticed that BelfrageTest sometimes failed locally and on CI
because a loop process would try to send data to Belfrage monitor, which
relies on checking a dial and unless the dials server is mocked, it
would fail. I also noticed that we never set any expectations on
Dials.Server and only stub it to make it return the values we need.

I thought that given the above, we could just start dials processes in
test environment and let them return the default values. We could
continue to stub them in tests that require non-default dial values.
This would mean that we wouldn't need to worry about dials in tests
where default values are fine.

This commit does the following:

* Renames `Dials.Server` to `Dials.LiveServer` and leaves the behaviour
only in `Server`, which `LiveServer` implements.
* Stubs `Dials.ServerMock` with `LiveServer` by default in tests, which
means that basically we'll have the actual implementation running in
tests, but can still stub dials if we need to.
* Ensures that dial processes are started by `Dials.Supervisor`.
`Dials.Poller` is still not started in test env, because I thought that
there was no point in having it running: dial values aren't going to
change in test env.
* Updated tests for `Dials.LiveServer` and `Dials.Supervisor`: I thought
that there's no real point in testing Elixir `Supervisor`'s behaviour,
like restarting dials on crashes, so removed those tests.

This change fixes the flaky test that I originally wanted to fix because
dials server will now be running.