Add a helper for stubbing dials in tests
Stubbing a dial involves calling the `transform` function on the dial
module, so I thought I'd add a helper to make it easier.

Values of dials are retrieved using a single function:
Belfrage.Dials.Server.state/1. This means two things:

* Stubbing a dial overwrites any previous stubs because that's the way
Mox works: if you stub a function it overwrites any previous stubs of
that function. This means that if we want to stub multiple dials we need
to add a stub that returns different values based on the name of the
dial. `stub_dials` takes care of that: it accepts a Keyword list of dial
names and values and stubs them all.
* The stubbed function will be called for any dial, not just the ones we
stub. To deal with the stubbed function delegates to `Dials.LiveServer`
if it's called with a dial that's not stubbed. In short, when stubbing
dials the stubbed ones will return the values that we want, other ones
will return default values.