Add a test helper module for caching
This new module replaces the `insert_cache_seed` helper, which is a bit
low level and not very user-friendly.

The new module also contains helpers for:

* Generating a cache key that is unique to the test, so you don't need
to worry that your simple cache key like "fresh" will clash with another
test that also uses that key.
* Generating a unique cache key: just a random value when you don't need
to know what it is, but just need to guarantee uniqueness.
* Clearing the cache: deleteing the :ets is a bit low level, so it's
good to have a reusable helper that encapsulates that (and also relies
on `Cachex` instead).

The idea of this module is to allow putting stuff into the cache and not
worry about clearing it or clashes with other tests.

I left the `start_test_cache` in place, but I don't think it's really
necessary to be able to start multiple `Cachex` processes: it should be
possible to just use a single one everywhere. I couldn't really
undersatnd what the `SupervisorTest` does it for, so left it for now.