Decouple Belfrage.RequestHash from BelfrageWeb.ResponseHeaders.Vary
As demonstrated by personalisation, the logic for caching responses in
Belfrage is not necessarily the same as for caching downstream, so it's
probably good not to have the corresponding modules depend on each other
and it's OK to have some duplication in them modules (e.g. ignore
`cookies` header in both independently).

I reckon this makes the logic a bit clearer: when generating a request
hash we ignore one set of headers, and when generating a `Vary` response
header - another.

I also removed `BelfrageWeb.ResponseHeaders.Vary.disallow_headers` which
was used in tests. Exclusion of `cookie` header was already tested
separately, so I don't think there is a point to also call
`disallow_headers` and test that the headers returned by it (only
`cookie` currently) are excluded: this is an example of future-proofing
that's neither necessary nor helpful.