Add default max_age to CachingHelper.put_into_cache
When the helper is used in tests, we simply want to put something into
cache and not worry about the mechanics of caching, e.g. in which case
responses are considered fresh or stale. This default value provides
that in the same way the existing default value for `cache_last_updated`
does.