Fetch fallback from cache in Belfrage.Processor
Moved the logic for fetching a cached resposne as a fallback to
`Belfrage.Processor` from `Belfrage.Cache.Fetch`.

I made this change in preparation for making fallback responses private
for personalised requests: I didn't want to add personalisation logic to
`Belfrage.Cache` because I belive that module shouldn't know anything
about that.

However, I think that this change makes sense in general and we should
let `Belfrage.Cache` take care of the mechanics of caching (like storing
and fetching, freshness, etc), while a layer higher up should contain
the logic that determines when to fetch a fallback (e.g. on which
errors, etc).