Don't kill loops in EndToEnd.VaryAllowHeadersTest
I've noticed that this test sometimes fails on CI with this error:

```
** (EXIT from #PID<0.6647.0>) exited in: GenServer.call({:via, Registry, {Belfrage.LoopsRegistry, {Belfrage.Loop, "SomeLoopAllowHeaders"}}}, :state, 750)
  ** (EXIT) no process: the process is not alive or there's no process currently associated with the given name, possibly because its application isn't started
```

I wonder if this is because we kill the loops before running the test
and `LoopSupervisor` doesn't manage to restart them before the test
executes a request. Restarting the loop is asynchronous, so there's no
guarantee that it'll actually happen.

I assume the test killed the loops to loose the state that loops
accumulate (like the number of failed requests that is used in
`CircuitBreaker`). I don't think it matters in this particular case
because the test is using a route spec that's not used anywhere else and
the requests don't fail.