Clear cache before running EndToEnd.MonitorEventsTest tests
I previously made it clear the cache after running tests, which I still
think is a good idea: tests shouldn't leave any side effects, however
there are other tests that make requests to `/200-ok-response` URL that
is used in this test and don't clear the cache afterwards, which is why
this test started to fail sometimes: in some cases there'd be an
unexpected cached response from another test and so we wouldn't call the
lambda.

We should probably update all end-to-end tests to clean-up after
themselves, but for now, I'm reverting to clearing the cache before
running these tests to stop them being flaky.