Make fallbacks for personalised requests private
When we receive an error from the origin in response to a personalised
request and serve a fallback, we don't want that fallback to be cached
downstream. The idea is that the error may only be happening to this
particular user and we don't want to cache the fallback for all signed
in users (which would happen because we vary the response on 'signed in'
HTTP header). To prevent this we mark the fallback response as
'private'.