Fetch early response for guest requests to personalised routes
If an unauthenticated user makes a request to a personalised route, we
should try to fetch an early response from cache. We only want to skip
that if the user is authenticated because we'd like to actually get a
personalised response from the origin.

We have to use `SessionState.authenticated?/1` for this purpose because
when an early response is fetched, request transformers haven't been
executed yet and so `Struct.user_session` is not populated.