Clear cache before running EndToEnd.PersonalisationTest tests
If there's a cached response in the cache (stored in another test while
processing a request to the URL that's used in this test) it can be used
instead of requesting the origin, so it's necessary to clear the cache
to ensure that it doesn't happen.

CachingHelper.clear_cache previously accepted the name of the cache to
clear, but this wasn't actually used anywhere, so removing that to be
able to use `clear_cache` as a setup function in tests (in which case it
receives current test context).