Fetch early response from cache if request is not personalised
We previously modified the logic for fetching early responses from cache
so that we skip it if a request is made to a personalised route and the
user is authenticated. However we later realised that we don't want to
skip it if e.g. the personalisation dial is off, but also, if the
request is not personalised for any other reason: we basically need to
perform all the checks in `Transformers.Personalisation` when deciding
whether to fetch an early response from cache or not.

To acheive this I made the following changes:

* Made `Struct.Private.personalised` attribute indicate whether the
request is personalised or not (i.e. it'll be true if the user is
authenticated, the route is personalised, the personalisation dial is
on, etc.). Before this change it indicated if the route was
personalised.
* Moved the logic that checks that a request is personalised from
`Transformers.Personalisation` to
`Belfrage.Personalisation.personalised_request?`
* Utilised that logic to set `Struct.Private.personalised` in
`Belfrage.Processor`, because it needs to run before we attempt to fetch
an early response from cache, i.e. before request transformers are
executed.

I also renamed `RouteSpec.Personalisation.enabled?` to `personalised?` for
clarity: that function checks whether a route is personalised or not and
`enabled?` could be read as "is personalisation enabled in general?"

Finally, I removed  redundant tests from
`Transformers.PersonalisationTest`:
* Some of the logic (various combiations of headers and cookies) is
already tested in `Authentication.SessionStateTest`, so just removed
those tests.
* Testing various invalid tokens belogs in `Authentication.TokenTest`,
so moved them.