Refactor BelfrageWeb.ResponseHeaders.VaryTest
Added helpers for calling Vary.add_header that return either the header
value as a string or the headers in that value as a list which make
testing easier.

Also:

* removed the shared structs because they don't save much space: many
tests need to modify them anyway
* updated the order of arguments in assertions: we normally assert that
a value returned by a function equals some value, not the other way
round.