Don't vary on x-id-oidc-signedin if personalisation is disabled
x-id-oidc-signedin header is allowed for personalised routes, but if
personalisation is disabled (e.g. the dial is off), we don't want to add
it to the Vary header that Belfrage serves because this makes GTM mark
the response as private. If personalisation is off, then the response is
public and we'd like to make GTM keep it as such.

We'd like to vary on x-id-oidc-signedin header if an unauthenticated
user visits a personalised route though: the response will be public and
if we don't vary on x-id-oidc-signedin it'll be served to authenticated
users by GTM.

In short, we don't want to vary on the header if an authenticated user
visits a personalised route: that would mean that personalisation is
disabled.