Replace Private.personalised with personalised_{route,request}
Before this change `Struct.Private.personalised` indicated if the
current request was personalised. It turns out that it's not enough to
know that and we need to know if the route is personalised (but e.g. the
user is not authenticated).

This commit replaces `Private.personalised` with two attributes:
`personalised_route` and `personalised_request`.

`personalised_route` comes from the route spec: when a route spec is
evaluated, it's set to true if personalisation is enabled for the route
in the current environment.

`personalised_request` is set as the `personalised` attribute before.
The difference is that it now relies on `personalised_route` attribute
which is set earlier instead of calling a function to check the current
`loop_id`.

I also moved the logic for transforming a route spec into a personalised
one to `Belfrage.Personalisation` to consolidate personalisation-related
logic in a single module. I think that it could be a good approach to
have feature-focused modules (or namespaces) that would provide
functionality to core parts of the system instead of scattering bits of
feature-specifc logic all over the system.