Don't vary on x-id-oidc-signedin if personalisation is disabled
I previously changed it so that we don't vary on x-id-oidc-signedin
header if user is authenticated and request is not personalised, however
that didn't work because if the request is not personalised, then we
don't poopulate `Struct.private.user_session`, so we won't know if the
user is authenticated or not.

After a bit more thought, I realised that there are four possible
scenarios:

* Unauthenticated user visits a personalised route and personalisation
is enabled: the response from GTM must be private because we don't want
the response to be cached downstream and served to an authenticated user
before it even reaches GTM, so we need to vary on `x-id-oidc-signedin`.
* Unauthenticated user visits a personalised route and personalisation
is disabled: the response is the same for all userse, so response from
GTM must be public and we shouldn't vary on `x-id-oidc-signedin`.
* Authenticated user visits a personalised route and personalisation is
enabled: the response from GTM must be private and so we need to vary on
`x-id-oidc-signedin`.
* Authenticated user visits a personalised route and personalisation is
disabled: the response is the same for all userse, so response from GTM
must be public and we shouldn't vary on `x-id-oidc-signedin`.

This basically means that we shouldn't vary on `x-id-oidc-signedin` when
personalisation is disabled, so that's the only thing we should check.

This commit includes an integration test that would have caught this
issue in the first place. It also renames `allowed_headers` to
`route_headers` in `BelfrageWeb.ResponseHeaders.Vary` to hopefully make
it less confusing: the function doesn't just return allowed headers, but
filters them as well.