[RESFRAME-4313] Use `private.personalised_request` field
In the `CacheDirective.call/1` function, use the
`private.personalised_request` field in the `%Belfrage.Struct`
in order determine whether or not a struct is personalised.

This also means the `set_stubs/1` function is no longer required
in the corresponding tests - so we can revert back to the use of
`set_webcore_ttl_multiplier/1` and `set_ttl_multiplier/1`,
both modified to use the `StubHelper.stub_dial/2` function.

The test structs have been removed as the belfrage `Struct`
includes a default value for `private.personalised_request`.