Use Macro.to_string when using platform name in metrics
I mistakenly removed `Macro.to_string` in
37378184b35d27f7fb79930114285684cb8ab64d because I thought that it
wasn't necessary, however, without it, `Private.platofrm` is
interpolated as e.g. `Elixir.Programmes` in the metric name because it's
a module. Using `Macro.to_string` ensures that it's just the name of the
module without the `Elixir.` prefix.

It'd be great to have tests for this (if we had, they'd cause this issue
in first place), but there's no quick way to add them because we track
metrics using `Belfrage.Event.record` and multple metrics are tracked
when a HTTP request is made, which makes it hard to test them.