Vary Lambda Alias not found errors based on env
This commit adds the passing through of the preview environment into
the Webcore response builder to allow us to return a different response
based on whether the preview environment is being used or not.
Preview environment on -> 404
Preview environment off -> 500
Only when the Lambda Alias error is :function_not_found, this addition
has no effect on other status codes as they are fine as they were.