Change Lambda function not found logging
The logic has changed so that we only a return a 404 for the preview
environment and a 500 for other envinronments when a :function_not_found
error is returned by a Lambda so it makes not sense to log a forced 404.
The message has also been changed to make it clearer that this is
related to the Lambda.