[RESFRAME-4371] Add and set S3 HTTP Client timeout to 1 second
In `Belfrage.generate_response/1`, a request is handled by
calling `Belfrage.Processor.perform_call/1`, which eventually
makes a lambda or http call - both of which have timeouts set
in the config: a `lambda_timeout` of `3_600` and a
`default_timeout` of `6_000`, respectively.

Later in `Belfrage.generate_response/1`,
`Belfrage.Processor.response_pipeline/1` is called, which
eventually calls `Belfrage.Processor.fetch_fallback_from_cache/1`.
This function eventually makes an S3 call when using a
`Distributed` cache, via `Belfrage.Clients.CCP.fetch/1`.
In this function the `default_timeout` is used when a HTTP request
is made to S3.

If we have a 10s Cowboy timeout, then a default timeout of 6s is
too long: timeouts of 6s to both the origin and S3 means that the
total time taken to finish both requests will be at least 12s -
more than 10s. Adding a timeout specifically for the S3 HTTP
request made in `Belfrage.Clients.CCP.fetch/1` means that the
total time taken for both requests will be 7s - less than 10s.