Don't reset loop counters in tests
When the "short" loop counter is reset, we send loop data to Belfrage
monitor, which involves checking the `:monitor_enabled` dial. Dials in
tests are mocked, so that dial needs to be stubbed before it can be
used, otherwise an error is raised by Mox. Resetting the loop counters
happens in the background, so that error can be really confusing when it
happens in a test that doesn't explicitly start any loops.

I previously attempted to fix this by making sure that dials are always
stubbed in dbe12f785ee300ca5d6604fa65821f29d36101f9. Unfortunatelly that
doesn't always work. Dials are initially stubbed in
`test/test_helper.exs` before any tests are run, however those stubs
then can temporarily stop working in tests that call
`Test.Support.Helper.mox`. It happens because the stubs are initally
added in the process that executes all tests (and that process gets
stored as the "global" process in `Mox`), however when
`Test.Support.Helper.mox` is called in a test that is not async, the
"global" process in Mox is overwritten with the PID of the current test
process and those inital stubs are lost. `Test.Support.Helper.mox` stubs
dials again as well and often that works fine, but there's still a delay
between updating of the "global" process in Mox and re-creation of the
stubs. If a dial is accessed during that delay, it won't be stubbed and
an error is raised. This sounds unlikely, but we reset the "short" loop
counter every 100ms in tests, so it does actually happen sometimes and
break tests.

I thought about an alternative fix for this issue and it seems like a
sensible solution would be not to reset loop counters in tests. We only
want that to happen in the actual tests for that functionality, but in
all other tests it's not necessary. Having a process that runs in the
background and can fail an unrelated test when its dependency is not
stubbed is confusing, so avoiding that sounds like a good thing. This
commit effectively disables reseting loop counters in test env by
setting the reset intervals to an arbitrary long value.

The tests that verify that counters are reset now modify the intervals
in the config and restart the loop so that those new values take effect.
They also set them back to original long values afterwards.