Don't call Utf8Sanitiser on query params
Invalid UTF chars in query params are currently caught by `Plug.Conn`,
Belfrage returns a 404 and doesn't attempt to make a request to the
origin as verified in the end-to-end tests in `NonUtf8QueryStringTest`

This means that `Utf8Sanitiser` will never be called when there's an
invalid UTF character in query params, so there's no point in ever
calling it. Another reason not to call it is that it can't handle query
params that are lists and requests with such parameters (e.g.
`/search?q[]=foo`) result in an exception.