Log invalid UTF8 chars in request path / query string
We'd like to understand what requests contain invalid UTF8 chars, so
adding a plug for that purpose. I'm adding a plug instead of logging
this in `BelfrageWeb.Router.error_handler` for the following reasons:

* Invalid UTF8 chars in request path don't currently result in an error
if the origin is not WebCore.
* For WebCore requests an error is raised when we attempt to generate
JSON and so it's not trivial to deduce from that error that the problem
is with invalid UTF8 in request path.
* I didn't want to polute the error handler, which is quite generic with
this very specific logic.

The plug will run before query params are fetched, which is when an
error is raised if they contain invalid UTF chars.