Remove `priv/static/dials.json`
This file was added in #463. I tried to understand from the PR and
comments what the reason for adding it was, but couldn't. It was
acknowledged that it's a duplication and a test was added to ensure that
it's in sync with `cosmos/dials.json`. I think that given that this file
is only used at compilation time and not at runtime, there's no need to
have it in `priv/static`.