Add Datalab Machine Recommendations dial
The SFV team are implementing Machine Driven Recommendations using a
service provided by Datalab. It is a service that returns clips that are
related to the entry clip the user is viewing based on machine learning.

They would like to be able to turn off that feature and this new dial
will serve that purpose. When the dial is enabled (which is the
default), Belfrage will pass a
`ctx-features=datalab_machine_recommendations=enabled` header to the
origin which will indicate that the feature is on. When the dial is off,
the header will be
`ctx-features=datalab_machine_recommendations=disabled`.

The header will only be set for specific routes: `SportVideos` and
`NewsVideos` for now. This is implemented using a new transformer that
can be added to the pipeline of the routes that require this feature.

I've opted to add a `Struct.private.features` field to implement this
because when we pass headers to Webcore we don't use the headers from
the incoming request. Plus I thought it could be confusing if a we had a
transformer that adds headers to the incoming request: you may end up
wondering whether a header came from the client or was added by
Belfrage.