Handle missing `request_received` timestamp on clean-up in LatencyMonitor
We've noticed these errors in logs during a load spike:

```
GenServer Belfrage.Metrics.LatencyMonitor terminating
** (CaseClauseError) no case clause matching: nil
    (stdlib 3.14) maps.erl:206: :maps.filter_1/2
    (stdlib 3.14) maps.erl:208: :maps.filter_1/2
    (stdlib 3.14) maps.erl:197: :maps.filter/2
    (belfrage 0.1.0) lib/belfrage/metrics/latency_monitor.ex:46: Belfrage.Metrics.LatencyMonitor.handle_info/2
```

The error means that when `LatencyMonitor` attempted to clean up old
events in its state and the following line was called:

```
:maps.filter(fn _request_id, times -> keep_request?(times, min_start_time) end, state)
```

`keep_request?/2` returned a `nil`, which could only happen if `times[:request_received]` was `nil`:

```
defp keep_request?(times, min_start_time) do
  times[:request_received] && times[:request_received] > min_start_time
end
```

I think what happened was:

* `request_receieved` was recorded for a request.
* More than 30s passed and `response_sent` still wasn't recorded for
that request because essentially system was slow under stress.
* The request was cleaned up.
* Some other event was eventually recorded for that request (e.g.
`origin_request_sent`).
* On the following clean-up `LatencyMonitor` came across the request, it
no longer had the `request_received` timestamp and that triggered the
error.

This commit fixes this by ensuring that `keep_request?/2` returns
`false` instead of `nil` in such scenario.

I considered whether we should not store any more events for a request
in such case (when `request_received` and potentially other timestamps
have already been cleaned up), but I think we still want to record them.
This is because in future (even though we currently don't) we'd still
like to be able to calculate internal latency, even if it's just based
on partial data: it's important to track internal latency when the
system is under stress.